requirements.txt
flask==2.3.3
flask-login==0.6.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
stripe==5.4.0
requests==2.31.0